New vision detection classes
