+++ src/software/sensor_fusion/BUILD
@@ -1,0 +1,11 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "detection",
+    srcs = ["detection.cpp"],
+    hdrs = ["detection.h"],
+    deps = [
+        "//software/geom",
+        "//software/util/time:timestamp",
+    ],
+)
